backend/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
Pillow
bcrypt
python-jose
elasticsearch